Made more Modular
- added separate Component for Card
- passed props to ServiceCard
- ensured background color of -
- ServiceCard also prop from now on so can easily change later on
- added api-host to .env as well
